avoid cast

+++ c_lines_bundle.c
@@ -91,7 +91,7 @@ int32 compare(const void *a, const void *b) {
 int32 median(int32 *array, int32 length) {
     qsort(array, length, sizeof(*array), compare);
     if ((length % 2) == 0) {
-        return (array[(int)((length/2) - 1)] + array[(int) (length/2)]) / 2.0;
+        return (array[(length/2) - 1] + array[length/2]) / 2.0;
     } else {
         return array[(int)(length/2)];
     }
